% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addlnorm.R
\name{addlnorm}
\alias{addlnorm}
\title{Overlay Exponential Density Function on Histogram}
\usage{
addlnorm(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A histogram of X overlayed with an exponential density function.
}
\description{
\code{add_exp} creates a histogram of \code{x} and overlays an exponential density function.
}
