% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayDensities.R
\name{addnorm}
\alias{addnorm}
\title{Overlay a Normal Density Function on Histogram}
\usage{
addnorm(
  x,
  myxlab = names(x),
  mytitle = "Histogram with normal curve",
  mynint = NULL
)
}
\arguments{
\item{x}{A numeric vector representing the data to be plotted.}

\item{myxlab}{What to label the x-axis.}

\item{mytitle}{What to title the graph.}

\item{mynint}{Number of bins.}
}
\value{
A histogram of x overlayed with an normal density function.
}
\description{
\code{addnorm} creates a histogram of \code{x} and overlays a normal density function.
Optional: Use myxlab to label the x-axis and mytitle to add a title.
mynint controls the number of bins
}
