% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{binom}
\alias{binom}
\title{Title}
\usage{
binom(k, prob, n, pi, lower.tail)
}
\arguments{
\item{lower.tail}{Boolean for finding the probability above (FALSE) or below
(TRUE) the inputted value (inclusive).}
}
\description{
Title
}
