% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{binompower}
\alias{binompower}
\title{Rejection Region for Binomial}
\usage{
binompower(LOS, n, prob1, alternative, prob2 = NULL)
}
\arguments{
\item{LOS}{A numeric value representing the level of significance}

\item{n}{A numeric value representing the sample size}

\item{prob1}{A numeric value representing the first probability}

\item{alternative}{A character value representing the alternative hypothesis}

\item{prob2}{A numeric value representing the second probability}
}
\value{
A plot of the binomial distribution with the rejection region highlighted.
}
\description{
\code{binompower} determines the rejection region corresponding to the level of
significance and the first probability and shows the binomial distribution
shading its corresponding region.
}
