% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{binomtest}
\alias{binomtest}
\title{Exact Binomial Test}
\usage{
binomtest(observed, n, hypothesized = NULL, alternative, conf.level = NULL)
}
\arguments{
\item{observed}{The observed number of successes or sample proportion (assumed
to be proportion if value less than one.)}

\item{n}{number of trials.}

\item{hypothesized}{hypothesized probability of success.}

\item{alternative}{Form of alternative ("less", "greater", or "two.sided").}

\item{conf.level}{Confidence level for a two-sided confidence interval.}
}
\value{
P-value along with a plot of the binomial distribution and/or
binomial confidence interval.
}
\description{
\code{binomtest} calculates performas an exact binomial test and graphs the
binomial distribution and/or binomial confidence interval.
}
