% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{boxplot}
\alias{boxplot}
\title{A box plot}
\usage{
boxplot(
  x,
  explanatory = NULL,
  main = "",
  xlab = "",
  ylab = substitute(explanatory)
)
}
\arguments{
\item{x}{Vector of numeric values to plot.}

\item{explanatory}{Optional second categorical variable to group by.}

\item{main}{Optional title for the plot}

\item{xlab}{Optional x-axis label for the plot}

\item{ylab}{Optional y-axis for the plot}
}
\value{
A box plot.
}
\description{
\code{boxplot} plots the given data in a box plot. If a second categorical variable
is given, the data is grouped by this variable.
}
