% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqrprob.R
\name{chisqrprob}
\alias{chisqrprob}
\title{Chi-Square Probability}
\usage{
chisqrprob(x_val, df)
}
\arguments{
\item{x_val}{the value of the chi-square statistic.}

\item{df}{the degrees of freedom.}
}
\value{
The upper tail probability for the chi-square distribution, and a
plot of the chi-square distribution with the statistic and more extreme
shaded.
}
\description{
\code{chisqrprob} returns the upper tail probability for the given chi-square
statistic and degress of freedom.
}
\examples{
chisqrprob(5, 3)
}
