% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_funcs_common_params}
\alias{.plot_funcs_common_params}
\title{Template documentation for plotting arguments}
\arguments{
\item{response}{Vector of numeric values to plot.}

\item{explanatory}{(optional) second categorical variable to group by.}

\item{main}{(optional) title for the plot.}

\item{xlab}{(optional) x-axis label for the plot.}

\item{ylab}{(optional) y-axis label for the plot. Only displayed when \code{explanatory} is provided.}
}
\description{
Template documentation for plotting arguments
}
\keyword{internal}
