% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Elephants}
\alias{Elephants}
\title{Elephant Walking Data}
\format{
\subsection{\code{Elephants}}{

A data frame with 56 rows and 2 columns:
\describe{
\item{Species}{What species this Elephant was.}
\item{Distance}{How many kilometers they walked per day.}
}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4944962/}
}
\usage{
Elephants
}
\description{
Researchers Holdgate et al. (2016) studied walking behavior of elephants in North American zoos to see
whether there is a difference in average distance traveled by African and Asian elephants in captivity.
They put GPS loggers on 33 African elephants and 23 Asian elephants, and measured the distance (in
kilometers) the elephants walked per day.
}
\keyword{datasets}
