% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Infant}
\alias{Infant}
\title{Infant Data}
\format{
\subsection{\code{Infant}}{

A data frame with 16 rows and 1 column:
\describe{
\item{choice}{Whether a baby selected the "helper" or "hinderer" toy.}
}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/18033298/}
}
\usage{
Infant
}
\description{
In a study reported in the November 2007 issue of Nature, researchers investigated whether infants take
into account an individual’s actions towards others in evaluating that individual as appealing or aversive,
perhaps laying for the foundation for social interaction (Hamlin, Wynn, and Bloom, 2007). In other
words, do children who aren’t even yet talking still form impressions as to someone’s friendliness based
on their actions? In one component of the study, 10-month-old infants were shown a “climber”
character (a piece of wood with “googly” eyes glued onto it) that could not make it up a hill in two
tries. Then the infants were shown two scenarios for the climber’s next try, one where the climber was
pushed to the top of the hill by another character (the “helper” toy) and one where the climber was
pushed back down the hill by another character (the “hinderer” toy). The infant was alternately shown
these two scenarios several times. Then the child was presented with both pieces of wood (the helper
and the hinderer characters) and asked to pick one to play with. Videos demonstrating this component
of the study can be found at \url{https://campuspress.yale.edu/infantlab/media/}.
}
\keyword{datasets}
