% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{invbinom}
\alias{invbinom}
\title{Inverse Binomial Probability}
\usage{
invbinom(alpha, n, prob, lower.tail)
}
\arguments{
\item{alpha}{The probability of interest.}

\item{n}{The number of trials.}

\item{prob}{The probability of success.}

\item{lower.tail}{Boolean for finding the probability above (FALSE) or below
(TRUE) the inputted value (inclusive)}
}
\value{
numeric which achieves at most the stated probability
}
\description{
Inverse Binomial Probability
}
