% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{invnorm}
\alias{invnorm}
\title{Inverse Normal Calculation}
\usage{
invnorm(prob1, mean = 0, sd = 1, direction)
}
\arguments{
\item{prob1}{probability to find normal quantile of.}

\item{mean}{mean of normal distribution.}

\item{sd}{standard deviation of normal distribution.}

\item{direction}{direction for probability calculation: "above", "below",
"outside", "between".}
}
\value{
a plot of the normal distribution with the quantile of the specified
probability highlighted.
}
\description{
Inverse Normal Calculation
}
