% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayDensities.R
\name{ISCAM.addexp}
\alias{ISCAM.addexp}
\title{Overlay an Exponential Density Function on Histogram}
\usage{
ISCAM.addexp(x)
}
\arguments{
\item{x}{A numeric vector representing the data to be plotted.}
}
\value{
A histogram of x overlayed with an exponential density function.
}
\description{
\code{addexp} creates a histogram of \code{x} and overlays an exponential density
function with \eqn{\lambda = \frac{1}{mean}}.
}
