% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayDensities.R
\name{ISCAM.addlnorm}
\alias{ISCAM.addlnorm}
\title{Overlay a Log Normal Density Function on Histogram}
\usage{
ISCAM.addlnorm(x)
}
\arguments{
\item{x}{A numeric vector representing the data to be plotted.}
}
\value{
A histogram of x overlayed with an log normal density function.
}
\description{
\code{addlnorm} creates a histogram of \code{x} and overlays a log normal density function.
}
