% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayDensities.R
\name{ISCAM.addt}
\alias{ISCAM.addt}
\title{Overlay a t Density Function on Histogram}
\usage{
ISCAM.addt(x, df)
}
\arguments{
\item{x}{A numeric vector representing the data to be plotted.}

\item{df}{A numeric value representing the degrees of freedom of \code{x}.}
}
\value{
A histogram of x overlayed with an t density function.
}
\description{
Overlay a t Density Function on Histogram
}
