% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{ISCAM.binomprob}
\alias{ISCAM.binomprob}
\title{Calculate Binomial Tail Probabilities}
\usage{
ISCAM.binomprob(k, n, prob, lower.tail)
}
\arguments{
\item{k}{number of successes of interest.}

\item{n}{number of trials.}

\item{prob}{success probability. Numeric between 0 & 1.}

\item{lower.tail}{Boolean for finding the probability above (FALSE) or below (TRUE) the inputted value (inclusive)}
}
\value{
The probability of the binomial distribution along with a graph of the distribution.
}
\description{
\code{binomprob} calculates the probability of the number of success of interest
using a binomial distribution and plots the distribution.
}
