% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ISCAM.boxplot}
\alias{ISCAM.boxplot}
\title{A box plot}
\usage{
ISCAM.boxplot(x, explanatory = NULL, names = NULL)
}
\arguments{
\item{x}{Vector of numeric values to plot.}

\item{explanatory}{Optional second categorical variable to group by.}

\item{names}{Vector of labels, horizontal and vertical axis respectively.}
}
\value{
A box plot.
}
\description{
\code{boxplot} plots the given data in a box plot. If a second categorical variable
is given, the data is grouped by this variable.
}
