% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{ISCAM.invt}
\alias{ISCAM.invt}
\title{Inverse T Calculation}
\usage{
ISCAM.invt(prob1, df, direction)
}
\arguments{
\item{prob1}{Desired probability.}

\item{df}{Degrees of freedom}

\item{direction}{direction for probability calculation: "above", "below",
"outside", "between".}
}
\value{
The t value for the specified probability.
}
\description{
\code{invt} calculates the t quantile of a specified probability.
}
