% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{ISCAM.normpower}
\alias{ISCAM.normpower}
\title{Rejection Region for Normal}
\usage{
ISCAM.normpower(LOS, n, prob1, alternative, prob2)
}
\arguments{
\item{LOS}{A numeric value representing the level of significance; 0 < \code{LOS}< 1}

\item{n}{A numeric value representing the sample size}

\item{prob1}{A numeric value representing the first probability}

\item{alternative}{A character value representing the alternative hypothesis}

\item{prob2}{A numeric value representing the second probability}
}
\value{
A plot of the normal distribution with the rejection region highlighted.
}
\description{
\code{normpower} determines the rejection region corresponding to the level of
significance and the first probability and shows the normal distribution
shading its corresponding region.
}
