% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_addexp.R
\name{iscam_addexp}
\alias{iscam_addexp}
\title{iscam_addexp Function}
\usage{
iscam_addexp(x)
}
\arguments{
\item{x}{a vector of numeric values.}
}
\description{
This function creates a histogram of the inputted variable and overlays 
an exponential density function with lambda = 1/mean.
}
\examples{
iscam_addexp(x)
}
\keyword{exponential}
