% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_addlnorm.R
\name{iscam_addlnorm}
\alias{iscam_addlnorm}
\title{iscam_addlnorm Function}
\usage{
iscam_addlnorm(x)
}
\arguments{
\item{x}{a vector of numeric values.}
}
\description{
This function creates a histogram of the inputted variable and overlays 
a log normal density function.
}
\examples{
iscam_addlnorm(x)
}
\keyword{lognormal}
