% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_addnorm.R
\name{iscam_addnorm}
\alias{iscam_addnorm}
\title{iscam_addnorm Function}
\usage{
iscam_addnorm(x)
}
\arguments{
\item{x}{a vector of numeric values.}
}
\description{
This function creates a histogram of the inputted variable and overlays 
a normal density function.
}
\examples{
iscam_addnorm(x)
}
\keyword{normal}
