% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_addt.R
\name{iscam_addt}
\alias{iscam_addt}
\title{iscam_addt Function}
\usage{
iscam_addt(x, df)
}
\arguments{
\item{x}{a vector of numeric values.}

\item{df}{degrees of freedom}
}
\description{
This function creates a histogram of the inputted variable and overlays
a t density function with df degrees of freedom.
}
\examples{
data("mtcars")
iscam_addt(mtcars$wt, 5)
}
\keyword{t}
