% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_binom.R
\name{iscam_binom}
\alias{iscam_binom}
\title{iscam_binom Function}
\usage{
iscam_binom(k, prob, n, pi, lower.tail)
}
\arguments{
\item{k}{number of successes}

\item{prob}{probability of interest}

\item{n}{number of trials (zero or more)}

\item{pi}{probability of success on each trial}

\item{lower.tail}{logical; if TRUE, probabilities are  P[X <= k], otherwise, P[X > k]}
}
\description{
This function calculates either a tail probability or the inverse cumulative 
probability depending on whether k or prob is passed a question mark.
}
\examples{
iscam_binom("?", 0.05, 20, 0.3, lower.tail = TRUE)
iscam_binom(10, "?", 20, 0.3, TRUE)
}
\keyword{binomial}
