% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_binomprob.R
\name{iscam_binomprob}
\alias{iscam_binomprob}
\title{iscam_binomprob Function}
\usage{
iscam_binomprob(k, n, prob, lower.tail)
}
\arguments{
\item{k}{number of successes of interest (must be an integer)}

\item{n}{number of trials (zero or more)}

\item{prob}{probability of success on each trial}

\item{lower.tail}{logical; a Boolean for finding the probability above (FALSE) or below (TRUE) the inputted value (inclusive)}
}
\description{
This function calculates tail probabilities from the binomial distribution.
}
\examples{
iscam_binomprob(20, 30, 0.5, lower.tail = TRUE)
iscam_binomprob(6, 20, 0.4, lower.tail = FALSE)
}
\keyword{binomial}
