% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_binomtest.R
\name{iscam_binomtest}
\alias{iscam_binomtest}
\title{iscam_binomtest Function}
\usage{
iscam_binomtest(observed, n, hypothesized, alternative, conf.level = NULL)
}
\arguments{
\item{observed}{number of successes of interest or sample proportion (assumed if value less than one)}

\item{n}{sample size}

\item{hypothesized}{probability of success on each trial}

\item{alternative}{form of alternative "less", "greater", or "two.sided"}

\item{conf.level}{NULL or a confidence level (one or more values) for a two-sided confidence interval}
}
\description{
This function performs an exact binomial test and graphs the binomial distribution and/or binomial confidence interval.
}
\examples{
iscam_binomtest(observed = 20, n = 35, hypothesized = 0.5, alternative = "less")
iscam_binomtest(10, 40, hypothesized = .50, alternative = "two.sided", conf.level = 0.90)
}
\keyword{binomial}
