% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_boxplot.R
\name{iscam_boxplot}
\alias{iscam_boxplot}
\title{iscam_boxplot Function}
\usage{
iscam_boxplot(x, explanatory = NULL, names = NULL)
}
\arguments{
\item{x}{a quantitative response variable.}

\item{explanatory}{a categorical explanatory variable.}

\item{names}{used to specify the horizontal and vertical axis labels, respectively.}
}
\description{
This function displays horizontal boxplot(s). 
Optional: A second, categorical variable can also be specified and values will be calculated separately for each group.
}
\examples{
iscam_boxplot(chickwts$weight)
iscam_boxplot(chickwts$weight, names = "weight")
iscam_boxplot(chickwts$weight, chickwts$feed, names = c("weight", "feed"))
}
\keyword{boxplot}
