% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_chisqprob.R
\name{iscam_chisqprob}
\alias{iscam_chisqprob}
\title{iscam_chisqprob Function}
\usage{
iscam_chisqprob(xval, df)
}
\arguments{
\item{xval}{x value}

\item{df}{degrees of freedom}
}
\description{
This function calculates the upper tail probability for the chi-square distribution
}
\examples{
iscam_chisqprob(10, 4)
}
\keyword{chi}
\keyword{square}
