% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_dotplot.R
\name{iscam_dotplot}
\alias{iscam_dotplot}
\title{iscam_dotplot Function}
\usage{
iscam_dotplot(response, explanatory = NULL, names = NULL)
}
\arguments{
\item{response}{a quantitative response variable.}

\item{explanatory}{a categorical explanatory variable.}

\item{names}{used to specify the horizontal and vertical axis labels, respectively.}
}
\description{
This function displays horizontal dotplot(s). 
Optional: A second, categorical variable can also be specified and values will be calculated separately for each group.
}
\examples{
iscam_dotplot(chickwts$weight)
iscam_dotplot(chickwts$weight, names = "weight")
iscam_dotplot(chickwts$weight, chickwts$feed, names = c("weight", "feed"))
}
\keyword{dotplot}
