% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_invbinom.R
\name{iscam_invbinom}
\alias{iscam_invbinom}
\title{iscam_invbinom Function}
\usage{
iscam_invbinom(alpha, n, prob, lower.tail)
}
\arguments{
\item{alpha}{alpha level}

\item{n}{sample size}

\item{prob}{despired probability of success}

\item{lower.tail}{upper tail (FALSE) or lower tail (TRUE)}
}
\description{
This function calculates the binomial quantile of a specified probability. The integer that achieves at most the stated probability will be returned.
}
\examples{
iscam_invbinom(.05, 15, .20, lower.tail = FALSE)
}
\keyword{binomial}
\keyword{inverse}
