% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_invnorm.R
\name{iscam_invnorm}
\alias{iscam_invnorm}
\title{iscam_invnorm Function}
\usage{
iscam_invnorm(prob1, mean = 0, sd = 1, direction)
}
\arguments{
\item{prob1}{desired probability}

\item{mean}{default = 0}

\item{sd}{default = 1}

\item{direction}{Specify whether you want this area to be "above", "below", or "outside" or "between"}
}
\description{
This function calculates the normal quantile of a specified probability.
}
\examples{
iscam_invnorm(prob1 = 0.25, direction = "above")
iscam_invnorm(.55, 15, 2, direction = "below")
iscam_invnorm(.40, direction = "outside")
iscam_invnorm(0.5, direction = "between")
}
\keyword{inverse}
\keyword{normal}
\keyword{probability}
