% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_invt.R
\name{iscam_invt}
\alias{iscam_invt}
\title{iscam_invt Function}
\usage{
iscam_invt(prob1, df, direction)
}
\arguments{
\item{prob1}{desired probability}

\item{df}{degrees of freedom}

\item{direction}{Specify whether you want this area to be "above", "below", or "outside" or "between"}
}
\description{
This function calculates the t quantile of a specified probability.
}
\examples{
iscam_invt(.35, 2, direction = "above")
iscam_invt(.4, 3, direction = "below")
iscam_invt(.45, 4, direction = "between")
iscam_invt(.55, 5, direction = "outside")
}
\keyword{inverse}
\keyword{probability}
\keyword{t}
