% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_normpower.R
\name{iscam_normpower}
\alias{iscam_normpower}
\title{iscam_normpower Function}
\usage{
iscam_normpower(LOS, n, prob1, alternative, prob2 = NULL, explain = FALSE)
}
\arguments{
\item{LOS}{level of significance}

\item{n}{sample size}

\item{prob1}{probability of success}

\item{alternative}{can be "less", "greater", or "two.sided"}

\item{prob2}{second probability of success}
}
\description{
This function determines the rejection region corresponding to the level of significance and the first probability and shows the second distribution shading its corresponding region.
}
\examples{
iscam_normpower(.05, 20, .5, "less", .7)
}
\keyword{normal}
\keyword{power}
