% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_onepropztest.R
\name{iscam_onepropztest}
\alias{iscam_onepropztest}
\title{iscam_onepropztest Function}
\usage{
iscam_onepropztest(observed, n, hypothesized = NULL,
  alternative = "two.sided", conf.level = NULL)
}
\arguments{
\item{observed}{number of success or sample proportion (assumed if value is less than 1)}

\item{n}{sample size}

\item{hypothesized}{hypothesized probability of success}

\item{alternative}{form of alternative hypothesis "less", "greater", or "two.sided"}

\item{conf.level}{confidence level(s) for a two-sided confidence interval}
}
\description{
This function calculates a one proportion z-test and/or interval.
}
\examples{
iscam_onepropztest(12, 15, .70, "less")
iscam_onepropztest(12, 15, .6, "greater")
iscam_onepropztest(12, 15, .5, "two.sided", 0.95)
iscam_onepropztest(12, 15, conf.level = c(.9, .95))
}
\keyword{one}
\keyword{proportion}
\keyword{test}
\keyword{z}
