% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_summary.R
\name{iscam_summary}
\alias{iscam_summary}
\title{iscam_summary Function}
\usage{
iscam_summary(x, explanatory = NULL, digits = 3)
}
\arguments{
\item{x}{quantitative variable}

\item{explanatory}{optional, categorical grouping variable}

\item{digits}{number of tick marks on x axis}
}
\description{
This function calculates the five number summary, mean, and standard deviation of the quantitative variable x.
}
\examples{
iscam_summary(chickwts$weight, chickwts$feed, digits = 4)
}
\keyword{summary}
