% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_tprob.R
\name{iscam_tprob}
\alias{iscam_tprob}
\title{iscam_tprob Function}
\usage{
iscam_tprob(xval, df, direction, xval2 = NULL)
}
\arguments{
\item{xval}{x value}

\item{df}{degrees of freedom}

\item{direction}{a string for finding the probability above ("above") or below ("below") the inputted value}

\item{xval2}{second observation necessary if "outside" or "between" is specified as the direction}
}
\description{
This function calculates tail probability for the t distribution.
}
\examples{
iscam_tprob(4, 5, "above")
}
\keyword{probability}
\keyword{t}
