% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMBinom.R
\name{ISCAMBinom}
\alias{ISCAMBinom}
\title{ISCAMBinom Function}
\usage{
ISCAMBinom(k, prob, n, pi, lower.tail)
}
\arguments{
\item{k}{number of successes}

\item{prob}{probability of interest}

\item{n}{number of trials (zero or more)}

\item{pi}{probability of success on each trial}

\item{lower.tail}{logical; if TRUE, probabilities are  P[X <= k], otherwise, P[X > k]}
}
\description{
This function calculates either a tail probability or the inverse cumulative 
probability depending on whether k or prob is passed a question mark.
}
\examples{
ISCAMBinom("?", 0.05, 20, 0.3, lower.tail = TRUE)
ISCAMBinom(10, "?", 20, 0.3, TRUE)
}
\keyword{binomial}
