% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{iscambinomnorm}
\alias{iscambinomnorm}
\title{Overlays Normal Approximation onto Binomial}
\usage{
iscambinomnorm(k, n, prob, direction)
}
\arguments{
\item{k}{number of successes of interest}

\item{n}{number of trials}

\item{prob}{success probability}

\item{direction}{"above", "below", or "two.sided"}
}
\value{
A plot of the binomial distribution overlayed with the normal approximation
}
\description{
\code{binomnorm} creates a binomial distribution of the given inputs and overlays a normal approximation.
}
\examples{
iscambinomnorm(k = 10, n = 20, prob = 0.5, direction = "two.sided")
}
