% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqprob.R
\name{iscamchisqprob}
\alias{iscamchisqprob}
\title{Chi-Square Probability}
\usage{
iscamchisqprob(xval, df, verbose = TRUE)
}
\arguments{
\item{xval}{the value of the chi-square statistic.}

\item{df}{the degrees of freedom.}

\item{verbose}{Logical, defaults to \code{TRUE}. Set to \code{FALSE} to suppress messages}
}
\value{
The upper tail probability for the chi-square distribution, and a
plot of the chi-square distribution with the statistic and more extreme
shaded.
}
\description{
\code{chisqrprob} returns the upper tail probability for the given chi-square
statistic and degrees of freedom.
}
\examples{
iscamchisqprob(5, 3)
}
