% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric.R
\name{iscamhyperprob}
\alias{iscamhyperprob}
\title{Hypergeometric p-value and Distribution}
\usage{
iscamhyperprob(k, total, succ, n, lower.tail, verbose = TRUE)
}
\arguments{
\item{k}{Number of successes of interest or difference in conditional proportions}

\item{total}{Total number of observations in the study}

\item{succ}{Overall number of successes}

\item{n}{Number of observations in group A}

\item{lower.tail}{Boolean for finding the probability above (FALSE) or
below (TRUE) the inputted value (inclusive)}

\item{verbose}{Logical, defaults to \code{TRUE}. Set to \code{FALSE} to suppress messages}
}
\value{
Tail probabilities from the hypergeometric distribution,
hypergeometric distribution with the observed statistic and more extreme
shaded.
}
\description{
Hypergeometric p-value and Distribution
}
\examples{
iscamhyperprob(1, 20, 5, 10, TRUE)
}
