% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{onesamplet}
\alias{onesamplet}
\title{One Sample T-Test}
\usage{
onesamplet(
  xbar,
  sd,
  n,
  hypothesized = 0,
  alternative = NULL,
  conf.level = NULL
)
}
\arguments{
\item{xbar}{Observed mean.}

\item{sd}{Observed standard deviation.}

\item{n}{Sample size.}

\item{hypothesized}{Hypothesized population mean.}

\item{alternative}{Alternative hypothesis: less/greater/two.sided}

\item{conf.level}{Confidence level.}
}
\value{
The t value, p value, and confidence interval.
}
\description{
\code{onesamplet} calculates a one sample t-test and/or interval from summary statistics.
It defaults to a hypothesized population mean of 0. You can optionally set an
alternative hypothesis and confidence level for a two-sided confidence interval.
}
