% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{tprob}
\alias{tprob}
\title{Tail Probability for t-distribution}
\usage{
tprob(xval, df, direction, xval2 = NULL)
}
\arguments{
\item{xval}{observed value.}

\item{df}{degrees of freedom.}

\item{direction}{direction for probability calculation, "above" or "below"; if
"outside" or "between" are used, a second larger observation, \code{xval2} must be
specified}

\item{xval2}{second observation value.}
}
\value{
The tail probability in the specified direction using the given
parameters.
}
\description{
Tail Probability for t-distribution
}
