% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{walmart}
\alias{walmart}
\title{Walmart Supercenter Data}
\format{
\subsection{\code{Walmart}}{

A data frame with 21 rows and 2 columns:
\describe{
\item{year}{The year.}
\item{stores}{The number of Walmart stores open in that year.}
\item{supercenters}{The number of Walmart superstores open in that year.}
}
}
}
\usage{
walmart
}
\description{
Walmart Supercenter Data
}
\keyword{datasets}
