% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addexp.R
\name{addexp}
\alias{addexp}
\title{Overlay Exponential Density Function on Histogram}
\usage{
addexp(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A histogram of X overlayed with an exponential density function.
}
\description{
\code{add_exp} creates a histogram of \code{x} and overlays an exponential density function with \eqn{\lambda = \frac{1}{mean}}.
}
\examples{
addexp(rchisq(30,3))
}
