% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddExp.R
\name{AddExp}
\alias{AddExp}
\title{AddExp Function}
\usage{
AddExp(x)
}
\arguments{
\item{x}{a vector of numeric values.}
}
\description{
This function creates a histogram of the inputted variable and overlays an exponential density function with lambda = 1/mean.
}
\examples{
AddExp()
}
\keyword{exponential}
