% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addlnorm.R
\name{addlnorm}
\alias{addlnorm}
\title{Overlay a Log Normal Density Function on Histogram}
\usage{
addlnorm(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A histogram of x overlayed with an log normal density function.
}
\description{
\code{addlnorm} creates a histogram of \code{x} and overlays a log normal density function.
}
