% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayDensities.R
\name{addnorm}
\alias{addnorm}
\title{Overlay a Normal Density Function on Histogram}
\usage{
addnorm(
  x,
  main = "Histogram with normal curve",
  xlab = deparse(substitute(x)),
  bins = NULL
)
}
\arguments{
\item{x}{A numeric vector representing the data to be plotted.}

\item{main}{What to title the graph.}

\item{xlab}{What to label the x-axis.}

\item{bins}{Number of bins for the histogram.}
}
\value{
A histogram of x overlayed with an normal density function.
}
\description{
\code{addnorm} creates a histogram of \code{x} and overlays a normal density function.
}
\examples{
set.seed(0)
x <- rnorm(100)
addnorm(x)
addnorm(x, main = "Your Active Title", xlab = "Normal Data", bins = 20)
}
