% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addt.R
\name{addt}
\alias{addt}
\title{Overlay a t Density Function on Histogram}
\usage{
addt(x, df)
}
\arguments{
\item{x}{A numeric vector.}

\item{df}{A numeric value.}
}
\value{
A histogram of x overlayed with an t density function.
}
\description{
Overlay a t Density Function on Histogram
}
