% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtnorm.R
\name{addtnorm}
\alias{addtnorm}
\title{Overlay a t Density Function & a Normal Density Function on Histogram}
\usage{
addtnorm(x, df)
}
\arguments{
\item{x}{A numeric vector representing the data to be plotted.}

\item{df}{A numeric value representing the degrees of freedom of \code{x}.}
}
\value{
A histogram of x overlayed with an t density function and a normal density function.
}
\description{
Overlay a t Density Function & a Normal Density Function on Histogram
}
