% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayDensities.R
\name{.add_density_common_params}
\alias{.add_density_common_params}
\title{Template documentation for plotting parameters}
\arguments{
\item{x}{A numeric vector representing the data to be plotted.}

\item{main}{What to title the graph.}

\item{xlab}{What to label the x-axis.}

\item{bins}{Number of bins for the histogram.}
}
\description{
Template documentation for plotting parameters
}
\keyword{internal}
