% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAM-package.R
\docType{package}
\name{ISCAM-package}
\alias{ISCAM}
\alias{ISCAM-package}
\title{ISCAM: Companion to the Book "Investigating Statistical Concepts, Applications, and Methods"}
\description{
Introductory statistics methods to accompany "Investigating Statistical Concepts, Applications, and Methods" (ISCAM) by Beth Chance & Allan Rossman (2024) \url{https://rossmanchance.com/iscam4/}. Tools to introduce statistical concepts with a focus on simulation approaches. Functions are verbose, designed to provide ample output for students to understand what each function does. Additionally, most functions are accompanied with plots. The package is designed to be used in an educational setting alongside the ISCAM textbook.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iscam4.github.io/ISCAM/}
  \item \url{https://github.com/ISCAM4/ISCAM}
  \item Report bugs at \url{https://github.com/ISCAM4/ISCAM/issues}
}

}
\author{
\strong{Maintainer}: Beth Chance \email{bchance@calpoly.edu} [copyright holder]

Authors:
\itemize{
  \item Visruth Srimath Kandali \email{public@visruth.com} (\href{https://orcid.org/0009-0005-9097-0688}{ORCID})
}

}
\keyword{internal}
