% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAM-package.R
\docType{package}
\name{ISCAM-package}
\alias{ISCAM}
\alias{ISCAM-package}
\title{ISCAM: Companion to the Book "Investigating Statistical Concepts, Applications, and Methods"}
\description{
Introductory statistics methods to accompany "Investigating Statistical Concepts, Applications, and Methods" by Beth Chance & Allan Rossman. Tools to introduce statistical concepts with a focus on simulation approaches. Functions are verbose, designed to provide ample output for students to understand what each function does. Additionally, most functions are accompanied with plots. The package is designed to be used in an educational setting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://visruthsk.github.io/ISCAM/}
}

}
\author{
\strong{Maintainer}: Beth Chance \email{bchance@calpoly.edu} [copyright holder]

Authors:
\itemize{
  \item Visruth Srimath Kandali \email{visruth@gmail.com} (\href{https://orcid.org/0009-0005-9097-0688}{ORCID})
}

}
\keyword{internal}
