% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam_addtnorm.R
\name{iscam_addtnorm}
\alias{iscam_addtnorm}
\title{iscam_addtnorm Function}
\usage{
iscam_addtnorm(x, df)
}
\arguments{
\item{x}{a vector of numeric values.}

\item{df}{degrees of freedom}
}
\description{
This function creates a histogram of the inputted variable and overlays a t density function with df degrees of freedom 
and a normal density function.
}
\examples{
iscam_addtnorm(x, 3)
}
\keyword{t}
