% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMAddLNorm.R
\name{ISCAMAddLNorm}
\alias{ISCAMAddLNorm}
\title{ISCAMAddLNorm Function}
\usage{
ISCAMAddLNorm(x)
}
\arguments{
\item{x}{a vector of numeric values.}
}
\description{
This function creates a histogram of the inputted variable and overlays 
a log normal density function.
}
\examples{
ISCAMAddLNorm(x)
}
\keyword{lognormal}
