% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMAddNorm.R
\name{ISCAMAddNorm}
\alias{ISCAMAddNorm}
\title{ISCAMAddNorm Function}
\usage{
ISCAMAddNorm(x)
}
\arguments{
\item{x}{a vector of numeric values.}
}
\description{
This function creates a histogram of the inputted variable and overlays 
a normal density function.
}
\examples{
ISCAMAddNorm(x)
}
\keyword{normal}
