% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMAddt.R
\name{ISCAMAddt}
\alias{ISCAMAddt}
\title{ISCAMAddt Function}
\usage{
ISCAMAddt(x, df)
}
\arguments{
\item{x}{a vector of numeric values.}
}
\description{
This function creates a histogram of the inputted variable and overlays 
a t density function with df degrees of freedom.
}
\examples{
ISCAMAddt(x, 5)
}
\keyword{t}
