% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMAddtNorm.R
\name{ISCAMAddtNorm}
\alias{ISCAMAddtNorm}
\title{ISCAMAddtNorm Function}
\usage{
ISCAMAddtNorm(x, df)
}
\arguments{
\item{x}{a vector of numeric values.}

\item{df}{degrees of freedom}
}
\description{
This function creates a histogram of the inputted variable and overlays a t density function with df degrees of freedom 
and a normal density function.
}
\examples{
ISCAMAddtNorm(x, 3)
}
\keyword{t}
