% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMBinomNorm.R
\name{ISCAMBinomNorm}
\alias{ISCAMBinomNorm}
\title{ISCAMBinomNorm Function}
\usage{
ISCAMBinomNorm(k, n, prob, direction)
}
\arguments{
\item{k}{number of successes of interest}

\item{n}{number of trials (zero or more)}

\item{prob}{probability of success on each trial}

\item{direction}{allows you to specify whether you want to find the probability 
"above" or "below" k or a symmetric "two.sided" probability}
}
\description{
This function illustrates the normal approximation to the binomial.
}
\examples{
ISCAMBinomNorm(20, 30, 0.5, direction = "above")
ISCAMBinomNorm(10, 55, 0.10, direction = "below")
ISCAMBinomNorm(13, 44, 0.40, direction = "two.sided")
}
\keyword{binomial}
\keyword{normal}
