% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMBinomPower.R
\name{ISCAMBinomPower}
\alias{ISCAMBinomPower}
\title{ISCAMBinomPower Function}
\usage{
ISCAMBinomPower(LOS, n, prob1, alternative, prob2 = NULL, explain = FALSE)
}
\arguments{
\item{LOS}{level of significance}

\item{n}{number of trials (zero or more)}

\item{prob1}{probability of success on each trial}

\item{alternative}{allows you to specify whether you want to find the probability 
"less" or "greater" or a symmetric "two.sided" probability}

\item{prob2}{NULL or a second probability}

\item{explain}{logical, default = FALSE. Set to TRUE to see type I and II error and power on the graph}
}
\description{
This function determines the rejection region corresponding to the level of 
significance and the first probability and shows the second distribution shading 
its corresponding region.
}
\examples{
ISCAMBinomPower(.05, 30, 0.5, alternative = "greater")
ISCAMBinomPower(.10, 55, 0.10, alternative = "less")
ISCAMBinomPower(.05, 20, 0.5, "two.sided", 0.6)
ISCAMBinomPower(.05, 20, 0.25, alternative = "greater",0.333, explain = T)
}
\keyword{binomial}
\keyword{power}
\keyword{region}
\keyword{rejection}
