% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMBoxplot.R
\name{ISCAMBoxplot}
\alias{ISCAMBoxplot}
\title{ISCAMBoxplot Function}
\usage{
ISCAMBoxplot(x, explanatory = NULL, names = NULL)
}
\arguments{
\item{x}{a quantitative response variable.}

\item{explanatory}{a categorical explanatory variable.}

\item{names}{used to specify the horizontal and vertical axis labels, respectively.}
}
\description{
This function displays horizontal boxplot(s). 
Optional: A second, categorical variable can also be specified and values will be calculated separately for each group.
}
\examples{
ISCAMBoxplot(chickwts$weight)
ISCAMBoxplot(chickwts$weight, names = "weight")
ISCAMBoxplot(chickwts$weight, chickwts$feed, names = c("weight", "feed"))
}
\keyword{boxplot}
