% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMChiSqProb.R
\name{ISCAMChiSqProb}
\alias{ISCAMChiSqProb}
\title{ISCAMChiSqProb Function}
\usage{
ISCAMChiSqProb(xval, df)
}
\arguments{
\item{xval}{x value}

\item{df}{degrees of freedom}
}
\description{
This function calculates the upper tail probability for the chi-square distribution
}
\examples{
ISCAMChiSqProb(10, 4)
}
\keyword{chi}
\keyword{square}
