% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMDotplot.R
\name{ISCAMDotplot}
\alias{ISCAMDotplot}
\title{ISCAMDotplot Function}
\usage{
ISCAMDotplot(response, explanatory = NULL, names = NULL)
}
\arguments{
\item{response}{a quantitative response variable.}

\item{explanatory}{a categorical explanatory variable.}

\item{names}{used to specify the horizontal and vertical axis labels, respectively.}
}
\description{
This function displays horizontal dotplot(s). 
Optional: A second, categorical variable can also be specified and values will be calculated separately for each group.
}
\examples{
ISCAMDotplot(chickwts$weight)
ISCAMDotplot(chickwts$weight, names = "weight")
ISCAMDotplot(chickwts$weight, chickwts$feed, names = c("weight", "feed"))
}
\keyword{dotplot}
