% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMInvNorm.R
\name{ISCAMInvNorm}
\alias{ISCAMInvNorm}
\title{ISCAMInvNorm Function}
\usage{
ISCAMInvNorm(prob1, mean = 0, sd = 1, direction)
}
\arguments{
\item{prob1}{desired probability}

\item{mean}{default = 0}

\item{sd}{default = 1}

\item{direction}{Specify whether you want this area to be "above", "below", or "outside" or "between"}
}
\description{
This function calculates the normal quantile of a specified probability.
}
\examples{
ISCAMInvNorm(prob1 = 0.25, direction = "above")
ISCAMInvNorm(.55, 15, 2, direction = "below")
ISCAMInvNorm(.40, direction = "outside")
ISCAMInvNorm(0.5, direction = "between")
}
\keyword{inverse}
\keyword{normal}
\keyword{probability}
