% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMInvt.R
\name{ISCAMInvt}
\alias{ISCAMInvt}
\title{ISCAMInvt Function}
\usage{
ISCAMInvt(prob1, df, direction)
}
\arguments{
\item{prob1}{desired probability}

\item{df}{degrees of freedom}

\item{direction}{Specify whether you want this area to be "above", "below", or "outside" or "between"}
}
\description{
This function calculates the t quantile of a specified probability.
}
\examples{
ISCAMInvt(.35, 2, direction = "above")
ISCAMInvt(.4, 3, direction = "below")
ISCAMInvt(.45, 4, direction = "between")
ISCAMInvt(.55, 5, direction = "outside")
}
\keyword{inverse}
\keyword{probability}
\keyword{t}
