% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMNormProb.R
\name{ISCAMNormProb}
\alias{ISCAMNormProb}
\title{ISCAMNormProb Function}
\usage{
ISCAMNormProb(xval, mean = 0, sd = 1, direction, label = NULL,
  xval2 = NULL, digits = 4)
}
\arguments{
\item{xval}{x value}

\item{mean}{default mean = 0}

\item{sd}{default standard deviation = 1}

\item{direction}{a String for finding the probability above "above" or below "below" the inputted value. If "outside" or "between" are specified, a second larger observation needs to be given at the end.}

\item{label}{It is highly recommended that you indicate a label for the horizontal axis, with the quotation marks e.g., "sample proportions"}

\item{xval2}{used if direction is specified as "outside" or "between"}

\item{digits}{number of tick marks on x axis}
}
\description{
This function calculates tail probability for the normal distribution.
}
\examples{
ISCAMNormProb(2.2, direction = "above", label = "sample proportions")
ISCAMNormProb(2, 5, 1.5, direction = "outside", label = "sample proportions", xval2 = 3)
}
\keyword{normal}
\keyword{probability}
\keyword{tail}
