% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMOneSamplet.R
\name{ISCAMOneSamplet}
\alias{ISCAMOneSamplet}
\title{ISCAMOneSamplet Function}
\usage{
ISCAMOneSamplet(xbar, sd, n, hypothesized = 0, alternative = NULL,
  conf.level = NULL)
}
\arguments{
\item{xbar}{observed mean}

\item{sd}{standard deviation}

\item{n}{sample size}

\item{hypothesized}{hypothesized population mean}

\item{alternative}{form of alternative hypothesis ("less", "greater", or "two.sided")}

\item{conf.level}{Input confidence level(s) for a two-sided confidence interval.}
}
\description{
This function calculates a one sample t-test and/or interval from summary statistics.
}
\examples{
ISCAMOneSamplet(70, 4, 25, hypothesized = 68, alternative = "greater")
}
\keyword{one}
\keyword{sample}
\keyword{t}
\keyword{test}
