% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscamsummary.R
\name{iscamsummary}
\alias{iscamsummary}
\title{Some Summary Statistics}
\usage{
iscamsummary(x, explanatory = NULL, digits = 3)
}
\arguments{
\item{x}{data to summarize.}

\item{explanatory}{optional explanatory variable to group by.}

\item{digits}{number of digits to round to.}
}
\value{
A table with some summary statistics of \code{x}.
}
\description{
\code{iscamsummary} calculates the five number summary, mean, and standard
deviation of the quantitative variable \code{x}. An optional second, categorical
variable can be specified and values will be calculated separately for
each group. The number of digits in output can also be specified.
}
