% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMTwoPropZTest.R
\name{ISCAMTwoPropZTest}
\alias{ISCAMTwoPropZTest}
\title{ISCAMTwoPropZTest Function}
\usage{
ISCAMTwoPropZTest(observed1, n1, observed2, n2, hypothesized = 0,
  alternative = NULL, conf.level = NULL, datatable = NULL)
}
\arguments{
\item{observed1}{observed number of successes or sample proportions (assumed if value is less than one)}

\item{n1}{sample size}

\item{observed2}{observed number of successes (assumed if value is less than one)}

\item{n2}{sample size}

\item{hypothesized}{hypothesized difference in probability of success/population proportions (default is zero)}

\item{alternative}{form of alternative hypothesis ("less", "greater", or "two.sided")}

\item{conf.level}{Input confidence level(s) for a two-sided confidence interval.}

\item{datatable}{Instead of entering the four individual numbers, you can pass in the two-way table of counts.  Be sure to label the datatable in this case.}
}
\description{
This function calculates a two proportion z-test and/or interval.
}
\examples{
ISCAMTwoPropZTest(6, 17, 12, 20, alternative = "less")
ISCAMTwoPropZTest(10, 20, 17, 22, conf.level = .95)
ISCAMTwoPropZTest(10, 20, 17, 22, conf.level = c(.90, .95))
}
\keyword{proportion}
\keyword{test}
\keyword{two}
\keyword{z}
