% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMTwoSamplet.R
\name{ISCAMTwoSamplet}
\alias{ISCAMTwoSamplet}
\title{ISCAMTwoSamplet Function}
\usage{
ISCAMTwoSamplet(x1, sd1, n1, x2, sd2, n2, hypothesized = 0,
  alternative = NULL, conf.level = 0)
}
\arguments{
\item{x1}{observed mean}

\item{sd1}{observed standard deviation}

\item{n1}{sample size}

\item{x2}{observed mean}

\item{sd2}{observed standard deviation}

\item{n2}{sample size}

\item{hypothesized}{hypothesized difference in population means (default = 0)}

\item{alternative}{form of alternative ("less", "greater", or "two.sided")}

\item{conf.level}{a confidence level for a two-sided confidence interval}
}
\description{
This function calculates a two sample t-test and/or interval from summary data.
}
\examples{
ISCAMTwoSamplet(97.25, 3.65, 8, 87.25, 9.60, 12, alternative = "less")
}
\keyword{sample}
\keyword{t}
\keyword{test}
\keyword{two}
