% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCAMTwoSamplet.R
\name{ISCAMTwoSamplet}
\alias{ISCAMTwoSamplet}
\title{ISCAMTwoSamplet Function}
\usage{
ISCAMTwoSamplet(observed1, n1, observed2, n2, hypothesized = 0,
  alternative = NULL, conf.level = NULL, datatable = NULL)
}
\arguments{
\item{n1}{sample size}

\item{n2}{sample size}

\item{hypothesized}{hypothesized difference in population means (default = 0)}

\item{alternative}{form of alternative ("less", "greater", or "two.sided")}

\item{conf.level}{a confidence level for a two-sided confidence interval}

\item{x1}{observed mean}

\item{sd1}{observed standard deviation}

\item{x2}{observed mean}

\item{sd2}{observed standard deviation}
}
\description{
This function calculates a two sample t-test and/or interval from summary data.
}
\examples{
ISCAMTwoSamplet(6, 1.7, 40, 8, 2, 60, alternative = "less")
}
\keyword{sample}
\keyword{t}
\keyword{test}
\keyword{two}
